"Set default condition to true in runCommand function
Updated the runCommand function in project-update.mjs to set the default value of the condition to 'true'. This change simplifies and improves the logic, making it easier to understand and control the flow of the script. It also ensures that the function will execute the command by default unless specified otherwise."