Replace prompt-sync with inquirer for project versioning
Replaced the `prompt-sync` package with `inquirer` package to provide a more intuitive console interface for version updating. Due to the change, `prompt-sync` was removed from `package.json` and `pnpm-lock.yaml` files and `inquirer` was added. In `project-update.mjs`, the script for updating version type was refactored to use `inquirer.prompt` instead of the previous `prompter`.