"Implemented projectBuilder and updated dependencies"
A new build script has been implemented using @ethang/project-builder. With this comes the transition of some npm dependencies to devDependencies, as they are not necessary for the application to run but are only required for the development process. The existing dependencies have been pruned and updated to reflect this change. The new build script ensures that project builds are more maintainable and easier to comprehend.