Consolidate execSync commands in project build
Modified the build script to consolidate the execSync commands into one. This was done to avoid repeatedly changing the current working directory during the project build process. Additionally, added 'stdio: "inherit"' to allow output to be logged to the console for easier debugging. This simplification of the build process offers improved maintainability.